refactor: separate rest client form
