--- src/types/syntax-highlighter.d.ts
@@ -1,4 +1,0 @@
-declare module 'react-syntax-highlighter/dist/esm/styles/prism' {
-  const styles: Record<string, Record<string, React.CSSProperties>>;
-  export = styles;
-}
